# Conflicts:
#	src/be/kdg/integration2/mvpglobal/MVPMain.java
#	src/be/kdg/integration2/mvpglobal/model/Screen.java
#	src/be/kdg/integration2/mvpglobal/model/Statistics.java
#	src/be/kdg/integration2/mvpglobal/utility/dbconnection/DBManager.java
#	src/be/kdg/integration2/mvpglobal/utility/dbconnection/mockdata.sql
#	src/be/kdg/integration2/mvpglobal/view/loginscreen/LoginScreenPresenter.java
#	src/be/kdg/integration2/mvpglobal/view/startscreen/StartScreenPresenter.java

+++ src/be/kdg/integration2/mvpglobal/utility/dbconnection/DDL.sql
@@ -59,7 +59,7 @@ CREATE TABLE IF NOT EXISTS PIECE_LOCATIONS (
 
 -- TABLE: MOVES
 CREATE TABLE IF NOT EXISTS MOVES (
-                                     moveID NUMERIC PRIMARY KEY,
+                                     moveID SERIAL PRIMARY KEY,
                                      sessionID SERIAL NOT NULL,
                                      move_start_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
                                      move_end_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
